+++ tests/Frontend/components/Footer.spec.js
@@ -21,7 +21,7 @@ describe('Footer Component', () => {
     moxios.uninstall();
   });
 
-  it('footer test with two urls', async () => {
+  it('footer test with two urls, no version', async () => {
     const view = mount(FooterComponent, {
       localVue,
       pinia: createTestingPinia(),
@@ -33,13 +33,14 @@ describe('Footer Component', () => {
     const settings = useSettingsStore();
     settings.settings = {
       legal_notice_url: 'https://legal.org',
-      privacy_policy_url: 'https://privacy.org'
+      privacy_policy_url: 'https://privacy.org',
+      version: null
     };
 
     await view.vm.$nextTick();
 
     // test if footer exists
-    const footer = view.findComponent({ ref: 'url_footer' });
+    const footer = view.findComponent({ ref: 'footer_container' });
     expect((footer).exists()).toBe(true);
 
     // test if text and | is shown and if correct links are inserted
@@ -50,9 +51,13 @@ describe('Footer Component', () => {
     expect(links.at(1).text()).toBe('app.footer.privacy_policy');
     expect(links.at(1).attributes('href')).toBe('https://privacy.org');
     expect(view.findComponent(RawText).exists()).toBeTruthy();
+
+    // test if version is not shown
+    expect(view.text()).not.toContain('app.version');
+    expect(view.text()).not.toContain('1.0.0');
   });
 
-  it('footer test with only legal notice', async () => {
+  it('footer test with only legal notice, no version', async () => {
     const view = mount(FooterComponent, {
       localVue,
       pinia: createTestingPinia(),
@@ -64,13 +69,15 @@ describe('Footer Component', () => {
     // set URLs for testing
     const settings = useSettingsStore();
     settings.settings = {
-      legal_notice_url: 'https://legal.org'
+      legal_notice_url: 'https://legal.org',
+      privacy_policy_url: '',
+      version: null
     };
 
     await view.vm.$nextTick();
 
     // test if footer exists
-    const footer = view.findComponent({ ref: 'url_footer' });
+    const footer = view.findComponent({ ref: 'footer_container' });
     expect((footer).exists()).toBe(true);
 
     // test if text and no | is shown and if correct links is inserted
@@ -79,9 +86,13 @@ describe('Footer Component', () => {
     expect(links.at(0).text()).toBe('app.footer.legal_notice');
     expect(links.at(0).attributes('href')).toBe('https://legal.org');
     expect(view.findComponent(RawText).exists()).toBeFalsy();
+
+    // test if version is not shown
+    expect(view.text()).not.toContain('app.version');
+    expect(view.text()).not.toContain('1.0.0');
   });
 
-  it('footer test with only privacy policy', async () => {
+  it('footer test with only privacy policy, no version', async () => {
     const view = mount(FooterComponent, {
       localVue,
       pinia: createTestingPinia(),
@@ -93,13 +104,15 @@ describe('Footer Component', () => {
     // set URLs for testing
     const settings = useSettingsStore();
     settings.settings = {
-      privacy_policy_url: 'https://privacy.org'
+      legal_notice_url: '',
+      privacy_policy_url: 'https://privacy.org',
+      version: null
     };
 
     await view.vm.$nextTick();
 
     // test if footer exists
-    const footer = view.findComponent({ ref: 'url_footer' });
+    const footer = view.findComponent({ ref: 'footer_container' });
     expect((footer).exists()).toBe(true);
 
     // test if text and no | is shown and if correct links is inserted
@@ -108,9 +121,13 @@ describe('Footer Component', () => {
     expect(links.at(0).text()).toBe('app.footer.privacy_policy');
     expect(links.at(0).attributes('href')).toBe('https://privacy.org');
     expect(view.findComponent(RawText).exists()).toBeFalsy();
+
+    // test if version is not shown
+    expect(view.text()).not.toContain('app.version');
+    expect(view.text()).not.toContain('1.0.0');
   });
 
-  it('footer test with no urls', function () {
+  it('footer test with no urls, no version', async () => {
     const view = mount(FooterComponent, {
       localVue,
       pinia: createTestingPinia(),
@@ -119,8 +136,160 @@ describe('Footer Component', () => {
       }
     });
 
+    // set URLs for testing
+    const settings = useSettingsStore();
+    settings.settings = {
+      legal_notice_url: '',
+      privacy_policy_url: '',
+      version: null
+    };
+
+    await view.vm.$nextTick();
+
     // test if no footer exists
-    const footer = view.findComponent({ ref: 'url_footer' });
+    const footer = view.findComponent({ ref: 'footer_container' });
     expect((footer).exists()).toBe(false);
+
+    // test if version is not shown
+    expect(view.text()).not.toContain('app.version');
+    expect(view.text()).not.toContain('1.0.0');
+  });
+
+  it('footer test with two urls and version', async () => {
+    const view = mount(FooterComponent, {
+      localVue,
+      pinia: createTestingPinia(),
+      mocks: {
+        $t: (key) => key
+      }
+    });
+
+    const settings = useSettingsStore();
+    settings.settings = {
+      legal_notice_url: 'https://legal.org',
+      privacy_policy_url: 'https://privacy.org',
+      version: '1.0.0'
+    };
+
+    await view.vm.$nextTick();
+
+    // test if footer exists
+    const footer = view.findComponent({ ref: 'footer_container' });
+    expect((footer).exists()).toBe(true);
+
+    // test if text and | is shown and if correct links are inserted
+    const links = footer.findAll('a');
+    expect(links.length).toBe(2);
+    expect(links.at(0).text()).toBe('app.footer.legal_notice');
+    expect(links.at(0).attributes('href')).toBe('https://legal.org');
+    expect(links.at(1).text()).toBe('app.footer.privacy_policy');
+    expect(links.at(1).attributes('href')).toBe('https://privacy.org');
+    expect(view.findComponent(RawText).exists()).toBeTruthy();
+
+    // test if version is shown
+    expect(view.text()).toContain('app.version');
+    expect(view.text()).toContain('1.0.0');
+  });
+
+  it('footer test with only legal notice and version', async () => {
+    const view = mount(FooterComponent, {
+      localVue,
+      pinia: createTestingPinia(),
+      mocks: {
+        $t: (key) => key
+      }
+    });
+
+    // set URLs for testing
+    const settings = useSettingsStore();
+    settings.settings = {
+      legal_notice_url: 'https://legal.org',
+      privacy_policy_url: '',
+      version: '1.0.0'
+    };
+
+    await view.vm.$nextTick();
+
+    // test if footer exists
+    const footer = view.findComponent({ ref: 'footer_container' });
+    expect((footer).exists()).toBe(true);
+
+    // test if text and no | is shown and if correct links is inserted
+    const links = footer.findAll('a');
+    expect(links.length).toBe(1);
+    expect(links.at(0).text()).toBe('app.footer.legal_notice');
+    expect(links.at(0).attributes('href')).toBe('https://legal.org');
+    expect(view.findComponent(RawText).exists()).toBeFalsy();
+
+    // test if version is shown
+    expect(view.text()).toContain('app.version');
+    expect(view.text()).toContain('1.0.0');
+  });
+
+  it('footer test with only privacy policy and version', async () => {
+    const view = mount(FooterComponent, {
+      localVue,
+      pinia: createTestingPinia(),
+      mocks: {
+        $t: (key) => key
+      }
+    });
+
+    // set URLs for testing
+    const settings = useSettingsStore();
+    settings.settings = {
+      legal_notice_url: '',
+      privacy_policy_url: 'https://privacy.org',
+      version: '1.0.0'
+    };
+
+    await view.vm.$nextTick();
+
+    // test if footer exists
+    const footer = view.findComponent({ ref: 'footer_container' });
+    expect((footer).exists()).toBe(true);
+
+    // test if text and no | is shown and if correct links is inserted
+    const links = footer.findAll('a');
+    expect(links.length).toBe(1);
+    expect(links.at(0).text()).toBe('app.footer.privacy_policy');
+    expect(links.at(0).attributes('href')).toBe('https://privacy.org');
+    expect(view.findComponent(RawText).exists()).toBeFalsy();
+
+    // test if version is shown
+    expect(view.text()).toContain('app.version');
+    expect(view.text()).toContain('1.0.0');
+  });
+
+  it('footer test with no urls, only version', async () => {
+    const view = mount(FooterComponent, {
+      localVue,
+      pinia: createTestingPinia(),
+      mocks: {
+        $t: (key) => key
+      }
+    });
+
+    // set URLs for testing
+    const settings = useSettingsStore();
+    settings.settings = {
+      legal_notice_url: '',
+      privacy_policy_url: '',
+      version: '1.0.0'
+    };
+
+    await view.vm.$nextTick();
+
+    // test if footer exists
+    const footer = view.findComponent({ ref: 'footer_container' });
+    expect((footer).exists()).toBe(true);
+
+    // test if no links are shown
+    const links = footer.findAll('a');
+    expect(links.length).toBe(0);
+
+    // test if version is shown
+    expect(view.text()).toContain('app.version');
+    expect(view.text()).toContain('1.0.0');
   });
 });